+++ de.metas.material/dispo-commons/src/main/java/de/metas/material/dispo/Candidate.java
@@ -63,6 +63,9 @@ public class Candidate
 	}
 
 	@NonNull
+	private final Integer clientId;
+	
+	@NonNull
 	private final Integer orgId;
 
 	@NonNull
@@ -106,6 +109,11 @@ public class Candidate
 	 */
 	private final ProductionCandidateDetail productionDetail;
 
+	/**
+	 * Used for additional infos if this candidate has the sub type {@link SubType#DISTRIBUTION}.
+	 */
+	private final DistributionCandidateDetail distributionDetail;
+	
 	/**
 	 * Used for additional infos if this candidate relates to particular demand
 	 */